CRAYSAT-1475: Update HTML builder for SAT 2.4 release
This commit adds configuration for the 2.4 release branch of docs-sat,
so that a 2.4 version is built and published to the online docs. This
commit modifies conf/sat.sh so that the existence of release/2.4 is
expected, adds hugo_prep and linkcheck containers for these branches,
and adds a new "language" to the Hugo config for 2.4.

Test Description:
* Ran the following:
** `bin/build.sh sat`
** `bin/test.sh sat`
** `bin/push.sh sat` (I commented out the line which does the push).